tests: convert vdsmdumpchains_test to pytest
Converting all tests in a module to pytest.

Namespace for exception values changed in
pytest.raises context manager to "value",
which used to be "exception".

Methods have been converted to functions
as the class is not needed anymore.

Converted to py3 which required only
removal of __future__ imports.

Checked with 2to3 utility and "python3.6
-m compileall -q vdsmdumpchains_test.py"

Signed-off-by: Roman Bednar <rbednar@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1870435
Change-Id: I06c52c46d3e436b8272758d9823ef91e89868a70