imagetickets: fix imageio text response parsing
Since 4.3 imageio returns text errors instead of json and was causing
imagetickets._read_content() to raise ImageDaemonError exception.

Parsing of response should be moved from _read_content() and done in
request() where we parse response based on content-type (and charset).

For raising errors message is decoded based on charset otherwise
returns bytes to caller.

get_ticket() is the only place where json has to be parsed so it's not
part of _request() anymore

New helper function for parsing charset from text/plain response has
been added and defaults to utf8. Also there are two new distinct
warning messages, one when additional paramenters from content-type
headers are missing and second when there is a parameter but it's not
"charset" - in both cases we default to utf8.

Signed-off-by: Roman Bednar <rbednar@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1858956
Change-Id: Ia1f5833791a9f0cf6bcf8226be883c6ee3fe15e6