tests: test adding a lease fails when index is full
Coverage report revealed missing test for a case when index runs out
of space.

The simplest test would be to keep adding leases using public method
LeasesVolume.add() until the index fills up. However since we use
temporary file backend to test leases volume this operation is slow.

To speed up the test we can add a memory backend which will serve
for testing purposes only and will be placed alongside the existing
backends in xleases.py so it is not hidden.

The memory backend is using io.BytesIO to simulate a file.

Test has a set timeout to prevent endless loop if something goes wrong
and the test would not work properly. Test fails if it times out.

Adding a lease logs info message which clogs the output when this test
runs (almost 4k entries). Using caplog fixture to silence the messages.

Signed-off-by: Roman Bednar <rbednar@redhat.com>
Change-Id: Ie802098fbb75ad971fcec33e90ba5083f31b4510