tests: add coverage for new index flows
When sanlock fails to write a resource when adding or removing a lease
the lookup fails with LeaseUpdating error.

Updating flag that we use in storage structure for a record in index
slot is the root cause of the related bug. The solution is to change
the flows so that the updating flag is not used. However removing it
completely is problematic because it can still be stored on the
underlying storage after updating vdsm to newer version.

So in case one domain contains host A using an old version another
host B using fixed version the presence of updating flag (set by
host A) must not cause failures on either host.

New flow has to be implemented for lease add, remove, and lookup
actions.

1) adding a lease
If there already is a record and is updating it should be used
and updating flag will reset. If no record is found for given lease id
we create a new entry which defaults to updating=False.

Also the order of actions needs to change as well:
1. Write sanlock resource
2. Update index

This will ensure the index is updated only if sanlock resource write
succeeded.

2) removing a lease
Removing the lease will unconditionally set the flag to False.

Order of actions is reversed than for adding:
1. Update index
2. Write sanlock resource

If the sanlock write fails we can just log the error instead of failing
as writing an empty record is enough for removing a lease.

3) lease lookup
For lookup we raise NoSuchLease if updating flag is set to True.

Signed-off-by: Roman Bednar <rbednar@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1902127
Change-Id: Icc9f75989328bef37490c4ff4c6fc60921350ea7