xleases: remove usage of updating flag during lease add/remove
Fixes the lease add, remove and lookup flows and related
documentation. Also enables tests with the new behavior.

For more details on the new flows see commit message of previous patch.

Changes to add/remove/lookup flows change the semantics of index, the
source of truth used to be the underlying storage format, now it's
the index. That means the index now no longer acts only as a cache
of the storage.

For example: if we remove a lease the index is updated first and
then sanlock can fail to clean up the resource on the storage. This is
enough to free up resources but also it means that we have an entry
on storage that is not valid and not present in the index buffer.
Rebuilding the index at this point using rebuild_index() would lead to
having orphaned leases in the index. This function needs to change in
future patches to something more suitable - e.g. iterating the index and
making sure all leases exists, adding them if needed.

Adding a lease works in opposite way, we only update the index if
sanlock succeeds to write a resource.

Signed-off-by: Roman Bednar <rbednar@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1902127
Change-Id: I67bf6a3140ab2d7ed9781808c85da7b23b53ecff