lvm: avoid exception for in use volume deactivation
If two VMs use the same ISO in block storage domain and one of the VMs
is terminated the teardownImage flow ends up raising an exception when
trying to deactivate logical volume for that image.

It is safe to ignore the deactivation attempt and instead of raising
this exception we can treat the volume as busy and just log that we
are ignoring a failed change of the volume.

Unfortunately lvm currently does not provide very reasonable exit codes
and so the only way to tell that a volume is busy is inspecting error
message of lvchange command.

Correct solution to this problem would be to implement reference
counting mechanism so we can keep track of volume usage. This
however requires more work and is planned to be added in future
versions.

Signed-off-by: Roman Bednar <rbednar@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1725915
Related-to: https://bugzilla.redhat.com/1536880
Change-Id: I88df42e45d6eb541385be4a0dfe64028a0bbef5f