volumemetadata: allow dumping incomplete metadata
While dumping volume metadata from storage we raise
MetaDataKeyNotFoundError if any of the values is missing.

This is not correct as it leaves out parts of the metadata which
might be present and valid.

This can be fixed by making VolumeMetadata more flexible,
more specifically VolumeMetadata.from_lines need to parse metadata
that can be incomplete.

Also it means the status of metadata has to be determined internally
(VolumeMetadata) based on detecting missing values. Until now we
assumed all values are present or KeyError is raised and status is set
externally (sc.VOL_STATUS_INVALID). That can not work if we want to
report incomplete data.

Signed-off-by: Roman Bednar <rbednar@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1870887
Change-Id: Iedc2021363c066ebf8a49bdc29cfbad4ea497a6e