volumemetadata: add parser method for validating storage data
Adding a parser function that can be either used to validate input
data for instantiating VolumeMetadata class using from_lines() or
it can be used to parse invalid or incomplete metadata without
instantiation of the class.

This method is flexible enough to be used for dumping invalid or
incomplete metadata by SD dump flow.

Signed-off-by: Roman Bednar <rbednar@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1870887
Change-Id: Iefa40f2b8467ceb9edbce07ec13840e01e11887c